logthrdestdrv: move logthrdestdrv to a separate directory under lib
