+++ lib/logthrdest/Makefile.am
@@ -1,0 +1,11 @@
+logthrdestincludedir = ${pkgincludedir}/logthrdest
+
+EXTRA_DIST += lib/logthrdest/CMakeLists.txt
+
+logthrdestinclude_HEADERS = \
+  lib/logthrdest/logthrdestdrv.h
+
+logthrdest_sources = \
+  lib/logthrdest/logthrdestdrv.c
+
+include lib/logthrdest/tests/Makefile.am
