+++ lib/logthrdest/logthrdestdrv.c
