+++ lib/logthrdest/tests/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_unit_test(CRITERION LIBTEST TARGET test_logthrdestdrv)
