+++ lib/logthrdest/tests/Makefile.am
@@ -1,0 +1,11 @@
+lib_logthrdest_tests_TESTS		= \
+	lib/logthrdest/tests/test_logthrdestdrv
+
+EXTRA_DIST += lib/logthrdest/tests/CMakeLists.txt
+
+check_PROGRAMS		+= ${lib_logthrdest_tests_TESTS}
+
+lib_logthrdest_tests_test_logthrdestdrv_CFLAGS	=	\
+	$(TEST_CFLAGS)
+lib_logthrdest_tests_test_logthrdestdrv_LDADD	=	\
+	$(TEST_LDADD)
