+++ lib/tests/CMakeLists.txt
@@ -7,7 +7,6 @@ add_unit_test(CRITERION TARGET test_hostname)
 add_unit_test(CRITERION LIBTEST TARGET test_rcptid)
 add_unit_test(CRITERION LIBTEST TARGET test_lexer)
 add_unit_test(CRITERION LIBTEST TARGET test_pragma)
-add_unit_test(CRITERION LIBTEST TARGET test_logthrdestdrv)
 add_unit_test(CRITERION TARGET test_lexer_block)
 add_unit_test(CRITERION TARGET test_str_format)
 add_unit_test(CRITERION TARGET test_str-utils)
